requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy==1.4.31
aiomysql
pymysql
python-jose[cryptography]
passlib[bcrypt]
